# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Programmatic access to the Weblens server",
        "title": "Weblens API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/",
    "paths": {
...
        "/media": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Get paginated media",
                "operationId": "GetMedia",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": false,
                        "description": "Include raw files",
                        "name": "raw",
                        "in": "query"
                    },
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": false,
                        "description": "Include hidden media",
                        "name": "hidden",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "createDate"
                        ],
                        "type": "string",
                        "default": "createDate",
                        "description": "Sort by field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page of medias to get",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of medias to get",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search only in given folders",
                        "name": "folderIds",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get only media with the provided ids",
                        "name": "mediaIds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media Batch",
                        "schema": {
                            "$ref": "#/definitions/MediaBatchInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/media/cleanup": {
            "post": {
                "security": [
                    {
                        "SessionAuth": [
                            "admin"
                        ]
                    },
                    {
                        "ApiKeyAuth": [
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Make sure all media is correctly synced with the file system",
                "operationId": "CleanupMedia",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/media/drop": {
            "post": {
                "security": [
                    {
                        "SessionAuth": [
                            "admin"
                        ]
                    },
                    {
                        "ApiKeyAuth": [
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "DANGEROUS. Drop all computed media and clear thumbnail in-memory and filesystem cache. Must be server owner.",
                "operationId": "DropMedia",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/media/visibility": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Set media visibility",
                "operationId": "SetMediaVisibility",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "description": "Set the media visibility",
                        "name": "hidden",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "MediaIds to change visibility of",
                        "name": "mediaIds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MediaIdsParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/media/{mediaId}.{extension}": {
            "get": {
                "produces": [
                    "image/*"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Get a media image bytes",
                "operationId": "GetMediaImage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media Id",
                        "name": "mediaId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Extension",
                        "name": "extension",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "thumbnail",
                            "fullres"
                        ],
                        "type": "string",
                        "description": "Image Quality",
                        "name": "quality",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "image bytes",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/media/{mediaId}/file": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Get file of media by id",
                "operationId": "GetMediaFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of media",
                        "name": "mediaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File info of file media was created from",
                        "schema": {
                            "$ref": "#/definitions/FileInfo"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/media/{mediaId}/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Get media info",
                "operationId": "GetMediaInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media Id",
                        "name": "mediaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media Info",
                        "schema": {
                            "$ref": "#/definitions/MediaInfo"
                        }
                    }
                }
            }
        },
        "/media/{mediaId}/liked": {
            "patch": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Like a media",
                "operationId": "SetMediaLiked",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of media",
                        "name": "mediaId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ShareId",
                        "name": "shareId",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Liked status to set",
                        "name": "liked",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/media/{mediaId}/video": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Stream a video",
                "operationId": "StreamVideo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of media",
                        "name": "mediaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "FileInfo": {
            "type": "object",
            "properties": {
                "childrenIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "contentId": {
                    "type": "string"
                },
                "currentId": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "hasRestoreMedia": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "isDir": {
                    "type": "boolean"
                },
                "modifiable": {
                    "type": "boolean"
                },
                "modifyTimestamp": {
                    "type": "integer"
                },
                "owner": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "pastFile": {
                    "type": "boolean"
                },
                "portablePath": {
                    "type": "string"
                },
                "shareId": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "MediaBatchInfo": {
            "type": "object",
            "properties": {
                "Media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MediaInfo"
                    }
                },
                "mediaCount": {
                    "type": "integer"
                }
            }
        },
        "MediaIdsParams": {
            "type": "object",
            "properties": {
                "mediaIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "MediaInfo": {
            "type": "object",
            "properties": {
                "contentId": {
                    "description": "Hash of the file content, to ensure that the same files don't get duplicated",
                    "type": "string"
                },
                "createDate": {
                    "type": "integer"
                },
                "duration": {
                    "description": "Total time, in milliseconds, of a video",
                    "type": "integer"
                },
                "enabled": {
                    "description": "If the media disabled. This can happen when the backing file(s) are deleted,\nbut the media stays behind because it can be re-used if needed.",
                    "type": "boolean"
                },
                "fileIds": {
                    "description": "Slices of files whos content hash to the contentId",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "height": {
                    "type": "integer"
                },
                "hidden": {
                    "description": "If the media is hidden from the timeline\nTODO - make this per user",
                    "type": "boolean"
                },
                "imported": {
                    "type": "boolean"
                },
                "likedBy": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mimeType": {
                    "description": "Mime-type key of the media",
                    "type": "string"
                },
                "owner": {
                    "description": "User who owns the file that resulted in this media being created",
                    "type": "string"
                },
                "pageCount": {
                    "description": "Number of pages (typically 1, 0 in not a valid page count)",
                    "type": "integer"
                },
                "recognitionTags": {
                    "description": "Tags from the ML image scan so searching for particular objects in the images can be done",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "width": {
                    "description": "Full-res image dimensions",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "scopes": {
                "admin": "Grants read and write access to privileged data"
            }
        },
        "SessionAuth": {
            "type": "apiKey",
            "name": "weblens-session-token",
            "in": "cookie"
        }
    }
}